+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/DefaultLabelService.java
@@ -4,7 +4,7 @@ import com.nhnacademy.marketgg.server.dto.request.LabelCreateRequest;
 import com.nhnacademy.marketgg.server.dto.response.LabelRetrieveResponse;
 import com.nhnacademy.marketgg.server.entity.Label;
 import com.nhnacademy.marketgg.server.exception.label.LabelNotFoundException;
-import com.nhnacademy.marketgg.server.repository.LabelRepository;
+import com.nhnacademy.marketgg.server.repository.label.LabelRepository;
 import com.nhnacademy.marketgg.server.service.LabelService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
