+++ src/test/java/com/nhnacademy/marketgg/server/service/impl/DefaultLabelServiceTest.java
@@ -13,7 +13,7 @@ import com.nhnacademy.marketgg.server.dto.request.LabelCreateRequest;
 import com.nhnacademy.marketgg.server.dto.response.LabelRetrieveResponse;
 import com.nhnacademy.marketgg.server.entity.Label;
 import com.nhnacademy.marketgg.server.exception.label.LabelNotFoundException;
-import com.nhnacademy.marketgg.server.repository.LabelRepository;
+import com.nhnacademy.marketgg.server.repository.label.LabelRepository;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
